split SPA into state and code files
